+++ NEWS.md
@@ -2,7 +2,9 @@
 
 * Postgres backend generates correct literals for Dates (#727).
 
-* `NA`s in `blob()`s are correctly translated to `NULL` (#983).
+* HANA backend correctly translates `as.character()` (#1027).
+
+* Correct translation for `median()` for SQL server (#1008).
 
 * `last()` is now correctly translated when no window frame is specified
   (@mgirlich, #1063).